Rank localized failures on p99 and open a sample from the table
The localization ratio was built from the max distance, which is a single vertex
and moves with one bad triangle; on real data the max runs to more than twice the
99th percentile. Rows now carry p99_dist and norm_p99, reported in the table, the
CSV and the Python API, and the ratio and its chart series use p99 instead.

Clicking a table row shows that sample on its own, which is the quickest way to
work down the ranking. Adds the first tests for CorrespondenceEvaluation, over
the in-memory entry point shared by the command line and Studio, including the
localized-defect case the metric exists to catch. Also fixes the Shape/Scalar
progress bar, which truncated its fractional progress the same way particle area
did.

+++ Studio/Analysis/ShapeScalarPanel.h
@@ -40,7 +40,7 @@ class ShapeScalarPanel : public QWidget {
 
   void run_clicked();
 
-  void handle_job_progress(int progress);
+  void handle_job_progress(double progress);
   void handle_job_complete();
 
  Q_SIGNALS:
